Split the remaining example tests into their own files.
