+++ src/gevent/tests/tests_that_dont_monkeypatch.txt
@@ -1,9 +1,23 @@
-test___example_servers.py
+test___config.py
+test___ident.py
+test___monitor.py
+test__ares_timeout.py
 test__backdoor.py
+test__close_backend_fd.py
+test__events.py
 test__example_echoserver.py
-test__example_udp_client.py
-test__getaddrinfo_import.py
 test__example_portforwarder.py
+test__example_udp_client.py
+test__example_wsgiserver.py
+test__example_wsgiserver_ssl.py
+test__example_webproxy.py
+test__examples.py
+test__getaddrinfo_import.py
+test__hub_join.py
+test__hub_join_timeout.py
+test__issue330.py
+test__iwait.py
+test__monkey_scope.py
 test__pywsgi.py
 test__server.py
 test__server_pywsgi.py
@@ -12,15 +26,3 @@ test__socket_dns6.py
 test__socket_errors.py
 test__socket_send_memoryview.py
 test__socket_timeout.py
-test__examples.py
-test__issue330.py
-test___ident.py
-test___config.py
-test___monitor.py
-test__events.py
-test__monkey_scope.py
-test__iwait.py
-test__ares_timeout.py
-test__close_backend_fd.py
-test__hub_join.py
-test__hub_join_timeout.py
